M4-Ph1 C1 extract clone_storage.py (F-201): move the 50-def clone-storage subsystem out of server.py
Verbatim move of the contiguous 50-def profile/clone-storage block
(_default_session_root .. _copy_clone_from_source + the 3 profile-selection
functions) into embedded/clone_storage.py. server.py imports the module and
delegates (spawn_browser -> clone_storage.resolve_profile_selection, the
app_lifespan background sweep, close_instance clone-dir release). cli.py's
storage reaches repoint through a new _clone_storage() accessor (_server()
stays for the non-storage stop/restart/kill-orphans env-guard). Public
functions drop the leading underscore; internal helpers keep it.

server.py 4425 -> 3389 LOC (grandfather cap ratcheted DOWN, no-grow).
clone_storage.py grandfathered at its actual 1057 LOC per the human gate
ruling 2026-07-12: the ~1024-line block is irreducibly >1000; the two-module
split and the partial move were both declined. pyproject per-file-ignores for
clone_storage.py carry the verbatim-moved suppressions (strict subset of
server.py's list), owner-tagged plan_M4ph1.

Pins: new tests/test_clone_storage.py (delegate-identity + spawn_browser
routing). test_profile_resolution.py and the M6 net repoint their patch
targets to the clone_storage home (semantics unchanged; predecessor bodies
kept byte-identical via import aliases). 677 passed / 1 skipped; all 6 gates
green.

Closes F-201. Partially closes F-106 (moved with a clear public name; the
retry/fallback shape is preserved, M6-pinned).
