+++ tests/test_profile_clone_excludes_cache.py
@@ -9,7 +9,7 @@ These tests pin the behavior: cache dirs are skipped, session state is kept.
 No browser required.
 """
 
-from stealth_chrome_devtools_mcp.embedded.server import (
+from stealth_chrome_devtools_mcp.embedded.clone_storage import (
     _copy_profile_delta,
     _profile_ignore_names,
 )
