+++ tests/test_profile_pid_check.py
@@ -7,15 +7,17 @@ server.py referenced ``psutil.Error`` at the handler but never imported ``psutil
 ``try`` body therefore masked itself with a NameError and crashed the caller.
 """
 
-from stealth_chrome_devtools_mcp.embedded import server
+from stealth_chrome_devtools_mcp.embedded import clone_storage
 
 
 def test_profile_pid_check_survives_os_error(tmp_path, monkeypatch):
     def _raise(_user_data_dir):
         raise OSError("simulated PID-lookup failure")
 
-    monkeypatch.setattr(server.process_cleanup, "_get_browser_pids_for_profile", _raise)
+    monkeypatch.setattr(
+        clone_storage.process_cleanup, "_get_browser_pids_for_profile", _raise
+    )
 
     # tmp_path carries no Chrome singleton markers, so the liveness heuristic must
     # fall back to False without raising.
-    assert server._profile_has_running_browser(tmp_path) is False
+    assert clone_storage._profile_has_running_browser(tmp_path) is False
